Add GUI theme support, ETA throttling, Win32 fixes
Introduce a GuiTheme setting (light/dark) across the app and wire it through config, GUI, and the Tk frontend. Added GuiTheme enum, config parsing/serialization, default and migration behavior, and tests/spec updates. Pass theme to the progress runner and extend the Tk GUI script to apply themes, add a Theme combobox in settings, and parse theme from session/progress args.

Reduce progress ETA update frequency by introducing scheduled/throttled ETA display updates in the Tcl progress UI and the Win32 progress dialog (ETA_UPDATE_MS) to cut excessive UI updates. Also add related state to track displayed ETA vs actual ETA.

Fix Win32 interop and dialog callbacks: adjust DialogCallback signature and dialog handler return types to Int64 (use i64 return values), add GetTickCount64 usage, and ensure consistent return values throughout Win32 UI code. Update minor constants and formatting. Updated specs to cover new gui_theme behavior.

+++ src/gui/win32/api.cr
@@ -88,7 +88,7 @@
 
       alias WinHWND = Void*
       alias WinBOOL = Int32
-      alias DialogCallback = Proc(Void*, UInt32, UInt64, UInt64, Int32)
+      alias DialogCallback = Proc(Void*, UInt32, UInt64, UInt64, Int64)
       alias BrowseCallback = Proc(Void*, UInt32, Int64, Int64, Int32)
 
       lib LibKernel32
